verificando se id existe

+++ src/main/java/com/vigjoaopaulo/gmail/com/controller/PerfilController.java
@@ -1,5 +1,6 @@
 package com.vigjoaopaulo.gmail.com.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,66 +18,85 @@ import org.springframework.web.bind.annotation.RestController;
 import com.vigjoaopaulo.gmail.com.model.Perfil;
 import com.vigjoaopaulo.gmail.com.repository.PerfilRepository;
 
-  @RestController
-  @RequestMapping
-  public class PerfilController {
+@RestController
+@RequestMapping
+public class PerfilController {
+
 	
-	  @Autowired
-		private PerfilRepository repository;
+	@Autowired
+	private PerfilRepository repository;
 
+	public boolean verificarPorId(Long id) {
+		return repository.existsById(id);
+	}
 
-		@GetMapping("/admim")
-		public List<Perfil> getPerfil(){
-			return repository.findAll();
-		}
+	public boolean verificarPorNome(Perfil perfil) {
+		return repository.existsByPerfil(perfil);
+	}
+
+	@GetMapping("/admim")
+	@ResponseStatus(HttpStatus.CREATED)
+	public List<Perfil> getPerfil() {
+		return repository.findAll();
+	}
+	
+
+	@PostMapping("/admim")
+	@ResponseStatus(HttpStatus.CREATED)
+	public String addPerfil(@RequestBody Perfil perfil) {
+		 
+		repository.findById(perfil.getId());
 		
-		@PostMapping("/admim")
-		public String addPerfil(@RequestBody Perfil perfil) {		
+		if(verificarPorId(null))
 
-			repository.save(perfil);
+		repository.save(perfil);
 
-			return "Inserido com sucesso id: ";
+		return "Inserido com sucesso id: ";
+
+	}
+
+	@GetMapping("/admim/{id}")
+	@ResponseStatus(HttpStatus.CREATED)
+	public String getId(@PathVariable Long id) {
+
+		if (verificarPorId(id) != true) {
+			return "Registro não encontrado...";
+		} else {
+			List<Perfil> lista = new ArrayList<>();
+			Perfil p = repository.findById(id).get();
+			p.getId();
+			p.getPerfil();
+			lista.add(p);
+
+			return "" + lista;
 		}
-		
-		@GetMapping("/admim/{id}")
-		public Perfil getId(@PathVariable Long id) {
-			
-			
-			
-				
-				return repository.findById(id).get();
-		}
-		
-		@DeleteMapping("/admim/{id}")
-		public String deletePerfil(@PathVariable Long id) {
+	}
+
+	@DeleteMapping("/admim/{id}")
+	@ResponseStatus(HttpStatus.CREATED)
+	public String deletePerfil(@PathVariable Long id) {
+
+		if (verificarPorId(id) != true) {
+			return "Registro não encontrado";
+		} else {
 			repository.deleteById(id);
-
 			return "Deletado com sucesso id: " + id;
 		}
-		
-		
-		@PutMapping("/admim/{id}")
-		@ResponseStatus(HttpStatus.CREATED)
-		public String updatePerfil(@PathVariable Long id,  @RequestBody Perfil perfil) {
 
-			Perfil newObj = repository.findById(id).get();
-			
-			newObj.setPerfil(perfil.getPerfil());
+	}
 
-		
-			repository.save(newObj);
+	@PutMapping("/admim/{id}")
+	@ResponseStatus(HttpStatus.CREATED)
+	public String updatePerfil(@PathVariable Long id, @RequestBody Perfil perfil) {
 
-			return "Atualizado com sucesso id: " + id;
+		Perfil newObj = repository.findById(id).get();
 
+		newObj.setPerfil(perfil.getPerfil());
 
-		}
-	
-	
-	
-	
-	
-	
-	
-	
-	
+		repository.save(newObj);
+
+		return "Atualizado com sucesso id: " + id;
+
+	}
+
 }
